reflecting post-secondary in Canada can be ISCED lvl 4
Trade schools, preliminary College years, and CEGEP (Quebec post-secondard schools) appear to be minimally ISCED level 4 in some cases. With the bar exclusively at 5, it will marginalize a very large portion of Quebec's post secondary school system known as CEGEP.
As well, organizations that handle trade are post-secondardy. Really, are we not talking about anyone talking post-secondary education pathways as 'in academia'?  It may be perceived as elitist as well and not really sensitive to the different practices worldwide.